backend/pyproject.toml
[tool.poetry]
name = "manim-generation"
version = "0.1.0"
description = "Visuals generator"
authors = [
    "Abhay Mishra <grabhaymishra@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
python-dotenv = ">=1.1.0,<2.0.0"
pydantic-settings = ">=2.9.1,<3.0.0"
supabase = "==2.12.0"
pydantic = ">=2.11.4,<3.0.0"
langchain-pinecone = ">=0.2.6,<0.3.0"
langchain-community = ">=0.3.23,<0.4.0"


python = ">=3.13,<3.14"
langchain-google-genai = "^2.1.4"
beautifulsoup4 = "^4.13.4"
lxml = "^5.4.0"
langchain-cohere = "^0.4.4"

[[tool.poetry.packages]]
include = "app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
